+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Sovereign Changlog
 
+## v0.20 - Feature Release
+* Added weapon table
+* Randomized starting weapon for player
+* Changed default creature XP
+
 ## v0.11 - Bugfix
 * Re-enabled damage.